+++ src/client/components/TenantsList/TenantsList.js
@@ -43,7 +43,7 @@ Template.TenantsList.onCreated( function(){
 Template.TenantsList.helpers({
     // whether the current user has the permission to see the list of tenants
     canList(){
-        const res = TenantsManager.isAllowed( 'pwix.tenants_manager.pub.list_all' );
+        const res = TenantsManager.isAllowed( 'pwix.tenants_manager.feat.list' );
         return res;
     },