+++ src/client/components/tm_entity_scoped_tab/tm_entity_scoped_tab.js
@@ -25,7 +25,7 @@ Template.tm_entity_scoped_tab.onCreated( function(){
 
     // check if the user is not allowed
     self.autorun(() => {
-        TenantsManager.isAllowed( 'pwix.tenants_manager.feat.roles', Meteor.userId(), Template.currentData().entity.get()._id ).then(( res ) => {
+        TenantsManager.isAllowed( 'pwix.tenants_manager.fn.edit', Meteor.userId(), Template.currentData().entity.get()._id ).then(( res ) => {
             self.TM.allowed.set( res );
         });
     });