+++ src/common/collections/tenants/server/functions.js
@@ -20,6 +20,45 @@ const logger = Logger.get();
 Tenants.s = {};
 
 /*
+ * @returns {Array} the array of the managers accounts ids
+ *  This is a private function (not protected by any permission) used in transformations when publishing
+ */
+Tenants.s._getManagers = async function( scope ){
+    let array = [];
+    const fetched = await Meteor.roleAssignment.find({ 'role._id': TenantsManager.configure().scopedManagerRole, scope: scope }).fetchAsync();
+    for( const it of fetched ){
+        const user = await Meteor.users.findOneAsync({ _id: it.user._id });
+        if( user ){
+            array.push( user );
+        } else {
+            // this may happen if a user has been authorized, then removed from the local account database without updating the roles
+            logger.warn( 'getManagers() user not found, but allowed by an assigned scoped role', it.user._id );
+            Tenants._errors = Tenants._errors || {};
+            Tenants._errors.manager_not_found = Tenants._errors.manager_not_found || {};
+            Tenants._errors.manager_not_found[it.user._id] = true;
+        }
+    }
+    return array;
+};
+
+/*
+ * @param {String} userId, may be null when called from common code on the server
+ * @returns {Array} the list of known tenants as objects { _id: <entity_id>, label: <closest_label> }
+ *  This is a private function (not protected by any permission)
+ *  Should rather use the corresponding reactive publication
+ */
+Tenants.s._getScopes = async function( userId ){
+    //check( userId, MatchOneOf( null, String ));
+    let result = [];
+    const entities = await Entities.collection.find().fetchAsync();
+    await Promise.allSettled( entities.map( async ( it ) => {
+        const records = await Records.collection.find({ entity: it._id }).fetchAsync();
+        result.push({ _id: it._id, label: Validity.closestByRecords( records ).record.label });
+    }));
+    return result;
+};
+
+/*
  * @param {String} pubName the publication name
  * @param {Object} tenantDoc
  * @param {Object} opts the options passed to the publish function
@@ -113,7 +152,7 @@ Tenants.s.applyUpdateTransforms = async function( fname, tenantDoc, opts={} ){
 Tenants.s.deleteTenant = async function( entity, userId ){
     check( entity, String );
     check( userId, Match.OneOf( null, String ));
-    if( !await TenantsManager.isAllowed( 'pwix.tenants_manager.fn.delete_tenant', userId, entity )){
+    if( !await TenantsManager.isAllowed( 'pwix.tenants_manager.fn.delete', userId, entity )){
         return null;
     }
     let result = {};
@@ -131,50 +170,8 @@ Tenants.s.deleteTenant = async function( entity, userId ){
 Tenants.s.getBy = async function( selector, userId ){
     //check( userId, MatchOneOf( null, String ));
     let result = {};
-    if( userId && !await TenantsManager.isAllowed( 'pwix.tenants_manager.fn.get_by', userId )){
-        return null;
-    }
     result.entities = await Entities.collection.find( selector ).fetchAsync();
     result.records = await Records.collection.find( selector ).fetchAsync();
-    return result;
-};
-
-/*
- * @returns {Array} the array of the managers accounts ids
- */
-Tenants.s.getManagers = async function( scope ){
-    let array = [];
-    const fetched = await Meteor.roleAssignment.find({ 'role._id': TenantsManager.configure().scopedManagerRole, scope: scope }).fetchAsync();
-    for( const it of fetched ){
-        const user = await Meteor.users.findOneAsync({ _id: it.user._id });
-        if( user ){
-            array.push( user );
-        } else {
-            // this may happen if a user has been authorized, then removed from the local account database without updating the roles
-            logger.warn( 'getManagers() user not found, but allowed by an assigned scoped role', it.user._id );
-            Tenants._errors = Tenants._errors || {};
-            Tenants._errors.manager_not_found = Tenants._errors.manager_not_found || {};
-            Tenants._errors.manager_not_found[it.user._id] = true;
-        }
-    }
-    return array;
-};
-
-/*
- * @param {String} userId, may be null when called from common code on the server
- * @returns {Array} the list of known tenants as objects { _id: <entity_id>, label: <closest_label> }
- */
-Tenants.s.getScopes = async function( userId ){
-    //check( userId, MatchOneOf( null, String ));
-    let result = [];
-    if( !await TenantsManager.isAllowed( 'pwix.tenants_manager.fn.get_scopes', userId )){
-        return null;
-    }
-    const entities = await Entities.collection.find().fetchAsync();
-    await Promise.allSettled( entities.map( async ( it ) => {
-        const records = await Records.collection.find({ entity: it._id }).fetchAsync();
-        result.push({ _id: it._id, label: Validity.closestByRecords( records ).record.label });
-    }));
     return result;
 };
 
@@ -198,44 +195,10 @@ Tenants.s.getTransformedEntities = async function(){
 Tenants.s.setManagers = async function( entity, userId ){
     check( entity, Object );
     check( userId, String );
-    if( !await TenantsManager.isAllowed( 'pwix.tenants_manager.fn.set_managers', userId, entity )){
+    if( !await TenantsManager.isAllowed( 'pwix.tenants_manager.fn.edit', userId, entity )){
         return null;
     }
 };
-
-/*
- * @param {Object} item the object read from the Entities collection
- * @returns {Object} this same entity object with its DYN sub-object
- */
-/*
-Tenants.s.transformEntity = async function( item ){
-    item.DYN = {
-        managers: [],
-        records: [],
-        closest: null
-    };
-    let promises = [];
-    promises.push( Tenants.s.getManagers( item._id ));
-    promises.push( Records.collection.find({ entity: item._id }).fetchAsync().then(( fetched ) => {
-        item.DYN.records = fetched;
-        item.DYN.closest = Validity.closestByRecords( fetched ).record;
-        return true;
-    }));
-    return Promise.allSettled( promises )
-        .then(() => {
-            // extend on option
-            const fn = TenantsManager.configure().serverAllExtend;
-            return fn ? fn( item ) : item;
-        })
-        .then(() => {
-            // make sure that each defined field appears in the returned item
-            // happens that clearing notes on server side does not publish the field 'notes' and seems that the previously 'notes' on the client is kept
-            // while publishing 'notes' as undefined rightly override (and erase) the previous notes on the client
-            Entities.s.addUndef( item );
-            return item;
-        });
-};
-*/
 
 /**
  * @param {String} name the publication name
@@ -270,7 +233,7 @@ Tenants.s.transformsUpdate = function(){
 Tenants.s.upsert = async function( entity, userId ){
     check( entity, Object );
     check( userId, String );
-    if( !await TenantsManager.isAllowed( 'pwix.tenants_manager.fn.upsert', userId, entity )){
+    if( !await TenantsManager.isAllowed( 'pwix.tenants_manager.fn.edit', userId, entity )){
         return null;
     }
     //logger.debug( 'upsert()', entity );