+++ src/common/collections/tenants/server/methods.js
@@ -10,10 +10,6 @@ Meteor.methods({
     },
 
     // update the managers of a tenant
-    async 'pwix.TenantsManager.m.Tenants.getScopes'(){
-        return await Tenants.s.getScopes( this.userId );
-    },
-
     async 'pwix.TenantsManager.m.Tenants.setManagers'( item ){
         return await Tenants.s.setManagers( item, this.userId );
     },